more graphics and a working top ten

+++ Assets/Scripts/Enemies/Scorable.cs
@@ -19,7 +19,7 @@ public class Scorable : MonoBehaviour {
 	void OnDeath() {
 		// When a scorable entity dies give player points and spawn a cool effect celebrating the fact
 		if(!GameApplication.WorldState.PlayerDead)
-			GameApplication.WorldState.Player.GetComponent<PlayerStatus>().AddScore(scoreValue, "KILL");
+			GameApplication.WorldState.Player.GetComponent<PlayerStatus>().AddScore((uint)scoreValue, "KILL");
 
 		GameObject s = Instantiate(scoreEffect, transform.position, Quaternion.identity) as GameObject;
 		s.GetComponent<TextEffect>().TextToDisplay = "+" + scoreValue.ToString();
